+++ lib/coinbase/client/models/broadcast_transfer_request.rb
@@ -14,22 +14,14 @@ require 'date'
 require 'time'
 
 module Coinbase::Client
-  # 
-  class FetchStakingRewards200Response
-    attr_accessor :data
-
-    # True if this list has another page of items after this one that can be fetched.
-    attr_accessor :has_more
-
-    # The page token to be used to fetch the next page.
-    attr_accessor :next_page
+  class BroadcastTransferRequest
+    # The hex-encoded signed payload of the transfer
+    attr_accessor :signed_payload
 
     # Attribute mapping from ruby-style variable name to JSON key.
     def self.attribute_map
       {
-        :'data' => :'data',
-        :'has_more' => :'has_more',
-        :'next_page' => :'next_page'
+        :'signed_payload' => :'signed_payload'
       }
     end
 
@@ -41,9 +33,7 @@ module Coinbase::Client
     # Attribute type mapping.
     def self.openapi_types
       {
-        :'data' => :'Array<StakingReward>',
-        :'has_more' => :'Boolean',
-        :'next_page' => :'String'
+        :'signed_payload' => :'String'
       }
     end
 
@@ -57,35 +47,21 @@ module Coinbase::Client
     # @param [Hash] attributes Model attributes in the form of hash
     def initialize(attributes = {})
       if (!attributes.is_a?(Hash))
-        fail ArgumentError, "The input argument (attributes) must be a hash in `Coinbase::Client::FetchStakingRewards200Response` initialize method"
+        fail ArgumentError, "The input argument (attributes) must be a hash in `Coinbase::Client::BroadcastTransferRequest` initialize method"
       end
 
       # check to see if the attribute exists and convert string to symbol for hash key
       attributes = attributes.each_with_object({}) { |(k, v), h|
         if (!self.class.attribute_map.key?(k.to_sym))
-          fail ArgumentError, "`#{k}` is not a valid attribute in `Coinbase::Client::FetchStakingRewards200Response`. Please check the name to make sure it's valid. List of attributes: " + self.class.attribute_map.keys.inspect
+          fail ArgumentError, "`#{k}` is not a valid attribute in `Coinbase::Client::BroadcastTransferRequest`. Please check the name to make sure it's valid. List of attributes: " + self.class.attribute_map.keys.inspect
         end
         h[k.to_sym] = v
       }
 
-      if attributes.key?(:'data')
-        if (value = attributes[:'data']).is_a?(Array)
-          self.data = value
-        end
+      if attributes.key?(:'signed_payload')
+        self.signed_payload = attributes[:'signed_payload']
       else
-        self.data = nil
-      end
-
-      if attributes.key?(:'has_more')
-        self.has_more = attributes[:'has_more']
-      else
-        self.has_more = nil
-      end
-
-      if attributes.key?(:'next_page')
-        self.next_page = attributes[:'next_page']
-      else
-        self.next_page = nil
+        self.signed_payload = nil
       end
     end
 
@@ -94,16 +70,8 @@ module Coinbase::Client
     def list_invalid_properties
       warn '[DEPRECATED] the `list_invalid_properties` method is obsolete'
       invalid_properties = Array.new
-      if @data.nil?
-        invalid_properties.push('invalid value for "data", data cannot be nil.')
-      end
-
-      if @has_more.nil?
-        invalid_properties.push('invalid value for "has_more", has_more cannot be nil.')
-      end
-
-      if @next_page.nil?
-        invalid_properties.push('invalid value for "next_page", next_page cannot be nil.')
+      if @signed_payload.nil?
+        invalid_properties.push('invalid value for "signed_payload", signed_payload cannot be nil.')
       end
 
       invalid_properties
@@ -113,9 +81,7 @@ module Coinbase::Client
     # @return true if the model is valid
     def valid?
       warn '[DEPRECATED] the `valid?` method is obsolete'
-      return false if @data.nil?
-      return false if @has_more.nil?
-      return false if @next_page.nil?
+      return false if @signed_payload.nil?
       true
     end
 
@@ -124,9 +90,7 @@ module Coinbase::Client
     def ==(o)
       return true if self.equal?(o)
       self.class == o.class &&
-          data == o.data &&
-          has_more == o.has_more &&
-          next_page == o.next_page
+          signed_payload == o.signed_payload
     end
 
     # @see the `==` method
@@ -138,7 +102,7 @@ module Coinbase::Client
     # Calculates hash code according to all attributes.
     # @return [Integer] Hash code
     def hash
-      [data, has_more, next_page].hash
+      [signed_payload].hash
     end
 
     # Builds the object from hash