+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.14.0] - 2025-01-14
+
+### Added
+- Add `skip_batching` option to `Wallet.transfer` to allow for lower latency gasless transfers.
+
+### Breaking
+- Removed support for EOL Ruby v2.7
+
 ## [0.13.0] - 2024-12-19
 
 ### Added